+++ tests/dummy/app/routes/search.js
@@ -1,0 +1,8 @@
+import Ember from 'ember';
+
+export default Ember.Route.extend({
+  model({ term }) {
+    return this.store.query('post', { filter: { term } });
+  },
+});
+